requirements.txt
aerich==0.7.2
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
dictdiffer==0.9.0
frozenlist==1.5.0
idna==3.10
iso8601==2.1.0
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.2.2
python-dotenv==1.0.1
pytz==2024.2
tomlkit==0.13.2
tortoise-orm==0.22.0
typing_extensions==4.12.2
yarl==1.17.2
